Drop support for Inline CMake toolchain
Initialy it was required when Kit does not correctly
pass compiller information for CMake. Now it does.
So quick and dyrty hack with inline toolchain overriding
does not required at all.

File toolchain per-project still keep: there is some amount
projects with complex toolchains for cross-building that can
and must be reused. Kit in this case only provide correct
compiler for correct parsing and so on.