CMake: Add QT_INSTALL_BINS to PATH on windows
Add QT_INSTALL_BINS to PATH on windows in the CMake RunConfiguration.

This should fix cmake applications using Qt not being able to get
started after they were built.

Unfortunately this patch is not a great solution: This should be configurable,
similar to the "Add libraries to LD_LIBRARY_PATH" checkbox we have for other
OSes and build systems. Such a change is unfortunately not acceptable at this
stage of the 4.5 branch though.

Task-number: QTCREATORBUG-19354
Change-Id: Ic2d568d965e41e3eb875fed8c2e5dd2ba15d2db1
Reviewed-by: Eike Ziller <eike.ziller@qt.io>