Fix broken "Clear CMake Configuration" action
We need to force reconfiguration instead of check one
when CMakeCache.txt isn't present. We just can't
check anything! There is not data to do it.

So, when we call "Clear CMake Configuration" and after that:

* Call "Run CMake" action, or
* Call "Rescan Project" action, or
* Just restart QtC

nothing will happens. And, when we try to build project,
it will fails, due to non-existing configuration at all.