Fix a possible crash when reaping a process with callbacks
It may happen that CMakeProcess destructor puts its internal
process into the reaper. After this, when the CMakeProcess
is already destructed, the internal process may still be working.
Since we have set the StdErrLineCallback lambda for the process
which accesses the private m_parser field, it may potentially
crash.

Solve it in a general way, so that whenever we put a process
to the reaper we reset its callbacks.

In addition, disconnect all the slots connected to process'
signals.

Fixes: QTCREATORBUG-26220
Change-Id: I63ea0e0ed6a62ef97ac695572f18014458489c5c
Reviewed-by: Cristian Adam <cristian.adam@qt.io>