cmake: Add option to disable subfolders
An option is added to change the default behavior that
would create subfolders inside source groups for files
based on their location on disk.

When disabling the option, files will be added directly
under their respective source group.

The virtual folders <build-folder> and <other-files>
is also omitted when the option is disabled.

Fixes: QTCREATORBUG-27432
Change-Id: Id78e178011c5299d4f7257bf855a5d791eebf91c
Reviewed-by: Cristian Adam <cristian.adam@qt.io>