CMakePM: Use cleaned paths for CMake parameter values
FilePath::fromUserInput will "clean" the paths, which on Windows means
that \\ will be replaced by /.

This is how CMake treats the paths internally as CMake paths.

This fixes the %{buildDir} macro which will expand to \\ on Windows.
Which causes issues with qtcsettings.cmake.

This way the expanded value of CMAKE_PROJECT_INCLUDE_BEFORE is not
displayed as red on Windows in the "Current Configuration" settings
page.

Change-Id: Ic26bf437de41ff3fb1c1b98d304ae84512cb0f1a
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>